Add files via upload
Completed the Videography section with a possible technique I read about in the pdf book that Steven Sent. Using ARTICLE to section out the video and date.  Could be wrong on that.  

I also completed the contact form, I know there are a lot of divs that don't make sense, they are intended to be for styling purposes only.